update cuda tool pathing
