+++ .gitignore
@@ -25,6 +25,8 @@ LocalPreferences.toml
 # Files generated by the build process
 libcxxwrap-julia*
 build*
+!deps/buildtools/
+!deps/buildtools/**
 wrapper/build
 
 legate_wrapper_install
