+++ deps/buildtools/cuda_tools.jl
@@ -1,0 +1,53 @@
+function resolve_custom_cuda(pkg_name::String)
+    cuda_enabled = Sys.which("nvcc") !== nothing
+    if !cuda_enabled
+        @warn "nvcc not found on PATH — building without CUDA. " *
+            "If your $(pkg_name) was built with CUDA, add nvcc to PATH and rebuild."
+    end
+    return cuda_enabled, nothing
+end
+
+function resolve_jll_cuda(jll_mod)
+    cuda_enabled = detect_jll_cuda_enabled(jll_mod)
+    cuda_root = try_get_cuda_sdk_jll_dir()
+    if isnothing(cuda_root) && cuda_enabled
+        @warn "CUDA_SDK_jll not found — cmake will search for system CUDA, which may " *
+            "not match the JLL. Add CUDA_SDK_jll to your environment for a reproducible build."
+    end
+    return cuda_enabled, cuda_root
+end
+
+function detect_jll_cuda_enabled(jll_mod)
+    cuda_val = get(jll_mod.host_platform.tags, "cuda", nothing)
+    return cuda_val !== nothing && cuda_val != "none"
+end
+
+# try/catch because CUDA_SDK_jll is a weakdep — may not be in the environment.
+function try_get_cuda_sdk_jll_dir()
+    try
+        Core.eval(Main, :(using CUDA_SDK_jll))
+        return joinpath(getfield(Main, :CUDA_SDK_jll).artifact_dir, "cuda")
+    catch
+        return nothing
+    end
+end
+
+function build_cuda_env(cuda_enabled::Bool, cuda_root)
+    # user-set env vars override auto-detection
+    if haskey(ENV, "LEGATE_WRAPPER_ENABLE_CUDA")
+        user_enable = uppercase(ENV["LEGATE_WRAPPER_ENABLE_CUDA"]) == "ON"
+        if user_enable && !cuda_enabled
+            @warn "LEGATE_WRAPPER_ENABLE_CUDA=ON overrides auto-detected cuda_enabled=false. " *
+                "The wrapper will link CUDA, but the underlying legate build may not — expect link/runtime errors if mismatched."
+        end
+        cuda_enabled = user_enable
+    end
+    if haskey(ENV, "CUDA_TOOLKIT_ROOT")
+        cuda_root = ENV["CUDA_TOOLKIT_ROOT"]
+    end
+
+    env = Dict{String,String}()
+    !cuda_enabled && (env["LEGATE_WRAPPER_ENABLE_CUDA"] = "OFF")
+    cuda_root !== nothing && (env["CUDA_TOOLKIT_ROOT"] = cuda_root)
+    return env
+end
