+++ scripts/build_cpp_wrapper.sh
@@ -37,10 +37,14 @@ if [[ ! -d "$INSTALL_DIR" ]]; then
     mkdir -p $INSTALL_DIR 
 fi
 
-NO_CUDA=${NO_CUDA:-OFF}
+# Optional env vars. Example:
+#   LEGATE_WRAPPER_ENABLE_CUDA=ON \
+#   CUDA_TOOLKIT_ROOT=/usr/local/cuda-13.0 \
+#       julia --project=. -e 'using Pkg; Pkg.build("Legate")'
+LEGATE_WRAPPER_ENABLE_CUDA=${LEGATE_WRAPPER_ENABLE_CUDA:-ON}
 CUDA_TOOLKIT_ROOT=${CUDA_TOOLKIT_ROOT:-}
 
-CUDA_ARGS=()
+CUDA_ARGS=("-DLEGATE_WRAPPER_ENABLE_CUDA=${LEGATE_WRAPPER_ENABLE_CUDA}")
 if [[ -n "$CUDA_TOOLKIT_ROOT" ]]; then
     CUDA_ARGS+=("-DCUDAToolkit_ROOT=${CUDA_TOOLKIT_ROOT}")
     CUDA_ARGS+=("-DCMAKE_LIBRARY_PATH=${CUDA_TOOLKIT_ROOT}/lib/stubs")
@@ -49,7 +53,6 @@ fi
 cmake -S $LEGATE_WRAPPER_SOURCE -B $BUILD_DIR \
     -D CMAKE_INSTALL_PREFIX=$INSTALL_DIR \
     -D BINARYBUILDER=OFF \
-    -D NO_CUDA=$NO_CUDA \
     -D CMAKE_PREFIX_PATH="$LEGATE_CMAKE_DIR;$LEGION_CMAKE_DIR;$REALM_CMAKE_DIR" \
     -D CMAKE_BUILD_TYPE=Release \
     "${CUDA_ARGS[@]}"
